move get.go to internal/app/get/*
